cloud bots utility file added , and the functions in it removed from the original bots
