+++ bots/sg_rules_delete_by_scope.py
@@ -28,7 +28,7 @@ Limitations: IPv6 is not supported
 
 """
 
-import re
+import bots_utils as utils
 
 PORT_TO = 'portTo'
 PORT_FROM = 'port'
@@ -37,25 +37,7 @@ SCOPE = 'scope'
 ALL_TRAFFIC_PORT = 0
 ALL_TRAFFIC_PROTOCOL = '-1'
 
-"""
-checks for ip validity ,else fix it to be right
-"""
 
-
-def verify_scope_is_cidr(rule):
-    ip = re.split('/|\.', str(rule[SCOPE]))  # break ip to blocks
-    rule[SCOPE] = ip[0] + '.' + ip[1] + '.' + ip[2] + '.' + ip[3] + '/' + ip[4]
-    pass
-
-
-"""
-returns a string of rule's id by scope,port,direction,etc.
-"""
-
-
-def stringify_rule(rule):
-    return 'rule: ' + rule[SCOPE] + ',' + str(rule[PORT_FROM]) + ',' + str(rule[PORT_TO]) + ',' + \
-           rule[PROTOCOL].lower() + ' '
 
 
 """
@@ -65,7 +47,7 @@ creates & removes the specified rules from a security group
 
 def delete_sg(sg, sg_id, rule, direction, text_output):
     # make sure that scope is in CIDR notation for example, 203.0.113.0/24
-    verify_scope_is_cidr(rule)
+    utils.verify_scope_is_cidr(rule)
 
     if direction == 'inbound':
         try:
@@ -76,7 +58,7 @@ def delete_sg(sg, sg_id, rule, direction, text_output):
                 GroupId=sg_id,
                 IpProtocol=rule[PROTOCOL].lower()
             )
-            text_output = text_output + stringify_rule(rule) + 'deleted successfully from sg : ' + str(sg_id) + '; '
+            text_output = text_output + utils.stringify_rule(rule) + 'deleted successfully from sg : ' + str(sg_id) + '; '
 
         except Exception as e:
             text_output = text_output + f'Error while trying to delete security group. Error: {e}'
